+++ jormungandr/src/explorer/graphql/mod.rs
@@ -1,3 +1,4 @@
+mod certificates;
 mod connections;
 mod error;
 mod scalars;
@@ -12,18 +13,18 @@ use super::indexing::{
 use super::persistent_sequence::PersistentSequence;
 use crate::blockcfg::{self, FragmentId, HeaderHash};
 use cardano_legacy_address::Addr as OldAddress;
+use certificates::*;
 use chain_impl_mockchain::certificate;
 use chain_impl_mockchain::key::BftLeaderId;
 use futures::executor::block_on;
 pub use juniper::http::GraphQLRequest;
 use juniper::{graphql_union, EmptyMutation, FieldResult, RootNode};
-use std::convert::TryFrom;
-use std::convert::TryInto;
+use std::convert::{TryFrom, TryInto};
 use std::str::FromStr;
 
 use self::scalars::{
-    BlockCount, ChainLength, EpochNumber, IndexCursor, NonZero, PoolId, PublicKey, Slot,
-    TimeOffsetSeconds, Value,
+    BlockCount, ChainLength, EpochNumber, ExternalProposalId, IndexCursor, NonZero, PoolId, Slot,
+    Value,
 };
 
 use crate::explorer::{ExplorerDB, Settings};
@@ -352,7 +353,7 @@ impl Transaction {
             .collect())
     }
 
-    pub fn certificate(&self, context: &Context) -> FieldResult<Option<Certificate>> {
+    pub fn certificate(&self, context: &Context) -> FieldResult<Option<certificates::Certificate>> {
         let transaction = self.get_contents(context)?;
         match transaction.certificate {
             Some(c) => Certificate::try_from(c).map(Some).map_err(|e| e.into()),
@@ -483,143 +484,6 @@ impl Address {
     }
 }
 
-/*--------------------------------------------*/
-/*------------------Certificates-------------*/
-/*------------------------------------------*/
-
-struct StakeDelegation {
-    delegation: certificate::StakeDelegation,
-}
-
-impl From<certificate::StakeDelegation> for StakeDelegation {
-    fn from(delegation: certificate::StakeDelegation) -> StakeDelegation {
-        StakeDelegation { delegation }
-    }
-}
-
-#[juniper::object(
-    Context = Context,
-)]
-impl StakeDelegation {
-    // FIXME: Maybe a new Account type would be better?
-    pub fn account(&self, context: &Context) -> FieldResult<Address> {
-        let discrimination = context.db.blockchain_config.discrimination;
-        self.delegation
-            .account_id
-            .to_single_account()
-            .ok_or(
-                // TODO: Multisig address?
-                ErrorKind::Unimplemented.into(),
-            )
-            .map(|single| {
-                chain_addr::Address(discrimination, chain_addr::Kind::Account(single.into()))
-            })
-            .map(|addr| Address::from(&ExplorerAddress::New(addr)))
-    }
-
-    pub fn pools(&self, context: &Context) -> Vec<Pool> {
-        use chain_impl_mockchain::account::DelegationType;
-        use std::iter::FromIterator as _;
-
-        match self.delegation.get_delegation_type() {
-            DelegationType::NonDelegated => vec![],
-            DelegationType::Full(id) => vec![Pool::from_valid_id(id.clone())],
-            DelegationType::Ratio(delegation_ratio) => Vec::from_iter(
-                delegation_ratio
-                    .pools()
-                    .iter()
-                    .cloned()
-                    .map(|(p, _)| Pool::from_valid_id(p)),
-            ),
-        }
-    }
-}
-
-#[derive(Clone)]
-struct PoolRegistration {
-    registration: certificate::PoolRegistration,
-}
-
-impl From<certificate::PoolRegistration> for PoolRegistration {
-    fn from(registration: certificate::PoolRegistration) -> PoolRegistration {
-        PoolRegistration { registration }
-    }
-}
-
-#[juniper::object(
-    Context = Context,
-)]
-impl PoolRegistration {
-    pub fn pool(&self, context: &Context) -> Pool {
-        Pool::from_valid_id(self.registration.to_id())
-    }
-
-    /// Beginning of validity for this pool, this is used
-    /// to keep track of the period of the expected key and the expiry
-    pub fn start_validity(&self) -> TimeOffsetSeconds {
-        self.registration.start_validity.into()
-    }
-
-    /// Management threshold for owners, this need to be <= #owners and > 0
-    pub fn management_threshold(&self) -> i32 {
-        // XXX: u8 fits in i32, but maybe some kind of custom scalar is better?
-        self.registration.management_threshold().into()
-    }
-
-    /// Owners of this pool
-    pub fn owners(&self) -> Vec<PublicKey> {
-        self.registration
-            .owners
-            .iter()
-            .map(PublicKey::from)
-            .collect()
-    }
-
-    pub fn operators(&self) -> Vec<PublicKey> {
-        self.registration
-            .operators
-            .iter()
-            .map(PublicKey::from)
-            .collect()
-    }
-
-    pub fn rewards(&self) -> TaxType {
-        TaxType(self.registration.rewards)
-    }
-
-    /// Reward account
-    pub fn reward_account(&self, context: &Context) -> Option<Address> {
-        use chain_impl_mockchain::transaction::AccountIdentifier;
-        let discrimination = context.db.blockchain_config.discrimination;
-
-        // FIXME: Move this transformation to a point earlier
-
-        self.registration
-            .reward_account
-            .clone()
-            .map(|acc_id| match acc_id {
-                AccountIdentifier::Single(d) => ExplorerAddress::New(chain_addr::Address(
-                    discrimination,
-                    chain_addr::Kind::Account(d.into()),
-                )),
-                AccountIdentifier::Multi(d) => {
-                    let mut bytes = [0u8; 32];
-                    bytes.copy_from_slice(&d.as_ref()[0..32]);
-                    ExplorerAddress::New(chain_addr::Address(
-                        discrimination,
-                        chain_addr::Kind::Multisig(bytes),
-                    ))
-                }
-            })
-            .map(|explorer_address| Address {
-                id: explorer_address,
-            })
-    }
-
-    // Genesis Praos keys
-    // pub keys: GenesisPraosLeader,
-}
-
 struct TaxType(chain_impl_mockchain::rewards::TaxType);
 
 #[juniper::object(
@@ -656,137 +520,12 @@ impl Ratio {
     }
 }
 
-struct OwnerStakeDelegation {
-    owner_stake_delegation: certificate::OwnerStakeDelegation,
-}
-
-impl From<certificate::OwnerStakeDelegation> for OwnerStakeDelegation {
-    fn from(owner_stake_delegation: certificate::OwnerStakeDelegation) -> OwnerStakeDelegation {
-        OwnerStakeDelegation {
-            owner_stake_delegation,
-        }
-    }
-}
 
 #[juniper::object(
     Context = Context,
 )]
-impl OwnerStakeDelegation {
-    fn pools(&self) -> Vec<Pool> {
-        use chain_impl_mockchain::account::DelegationType;
-        use std::iter::FromIterator as _;
-
-        match self.owner_stake_delegation.get_delegation_type() {
-            DelegationType::NonDelegated => vec![],
-            DelegationType::Full(id) => vec![Pool::from_valid_id(id.clone())],
-            DelegationType::Ratio(delegation_ratio) => Vec::from_iter(
-                delegation_ratio
-                    .pools()
-                    .iter()
-                    .cloned()
-                    .map(|(p, _)| Pool::from_valid_id(p)),
-            ),
-        }
     }
 }
-
-/// Retirement info for a pool
-struct PoolRetirement {
-    pool_retirement: certificate::PoolRetirement,
-}
-
-impl From<certificate::PoolRetirement> for PoolRetirement {
-    fn from(pool_retirement: certificate::PoolRetirement) -> PoolRetirement {
-        PoolRetirement { pool_retirement }
-    }
-}
-
-#[juniper::object(
-    Context = Context,
-)]
-impl PoolRetirement {
-    pub fn pool_id(&self) -> PoolId {
-        PoolId(format!("{}", self.pool_retirement.pool_id))
-    }
-
-    pub fn retirement_time(&self) -> TimeOffsetSeconds {
-        self.pool_retirement.retirement_time.into()
-    }
-}
-
-struct PoolUpdate {
-    pool_update: certificate::PoolUpdate,
-}
-
-impl From<certificate::PoolUpdate> for PoolUpdate {
-    fn from(pool_update: certificate::PoolUpdate) -> PoolUpdate {
-        PoolUpdate { pool_update }
-    }
-}
-
-#[juniper::object(
-    Context = Context,
-)]
-impl PoolUpdate {
-    pub fn pool_id(&self) -> PoolId {
-        PoolId(format!("{}", self.pool_update.pool_id))
-    }
-
-    pub fn start_validity(&self) -> TimeOffsetSeconds {
-        self.pool_update.new_pool_reg.start_validity.into()
-    }
-
-    // TODO: Previous keys?
-    // TODO: Updated keys?
-}
-
-// TODO can we use jormungandr-lib Certificate ?
-enum Certificate {
-    StakeDelegation(StakeDelegation),
-    OwnerStakeDelegation(OwnerStakeDelegation),
-    PoolRegistration(PoolRegistration),
-    PoolRetirement(PoolRetirement),
-    PoolUpdate(PoolUpdate),
-}
-
-impl TryFrom<chain_impl_mockchain::certificate::Certificate> for Certificate {
-    type Error = error::Error;
-    fn try_from(
-        original: chain_impl_mockchain::certificate::Certificate,
-    ) -> Result<Certificate, Self::Error> {
-        match original {
-            certificate::Certificate::StakeDelegation(c) => {
-                Ok(Certificate::StakeDelegation(StakeDelegation::from(c)))
-            }
-            certificate::Certificate::OwnerStakeDelegation(c) => Ok(
-                Certificate::OwnerStakeDelegation(OwnerStakeDelegation::from(c)),
-            ),
-            certificate::Certificate::PoolRegistration(c) => {
-                Ok(Certificate::PoolRegistration(PoolRegistration::from(c)))
-            }
-            certificate::Certificate::PoolRetirement(c) => {
-                Ok(Certificate::PoolRetirement(PoolRetirement::from(c)))
-            }
-            certificate::Certificate::PoolUpdate(c) => {
-                Ok(Certificate::PoolUpdate(PoolUpdate::from(c)))
-            }
-            certificate::Certificate::VotePlan(_c) => todo!("Vote plans are not yet supported"),
-            certificate::Certificate::VoteCast(_c) => todo!("Vote casts are not yet supported"),
-        }
-    }
-}
-
-graphql_union!(Certificate: Context |&self| {
-    // the left hand side of the `instance_resolvers` match-like structure is the one
-    // that's used to match in the graphql query with the `__typename` field
-    instance_resolvers: |_| {
-        &StakeDelegation => match *self { Certificate::StakeDelegation(ref c) => Some(c), _ => None },
-        &OwnerStakeDelegation => match *self { Certificate::OwnerStakeDelegation(ref c) => Some(c), _ => None },
-        &PoolRegistration => match *self { Certificate::PoolRegistration(ref c) => Some(c), _ => None },
-        &PoolUpdate => match *self { Certificate::PoolUpdate(ref c) => Some(c), _ => None},
-        &PoolRetirement => match *self { Certificate::PoolRetirement(ref c) => Some(c), _ => None}
-    }
-});
 
 #[derive(Clone)]
 pub struct Pool {
